Move About struct from library into Engine
